The core is fully functional
It is also heavily untested. We don't know if the results are indeed the expected ones. The only confirmation that we have, is that it works properly, and that is because of its complexity, (if something is wrong, the whole system will go down), and because of the "normality" of the output.